+++ app/src/main/java/com/dofury/moneycycle/dto/MoneyLog.kt
@@ -1,3 +1,3 @@
 package com.dofury.moneycycle.dto
 
-data class MoneyLog(var uid: Int,var charge: Int, var sign: Boolean, var category: String, var date: String,var is_server: Boolean)
+data class MoneyLog(var uid: Int,var charge: Long, var sign: Boolean, var category: String, var date: String,var is_server: Boolean)
